[refactor] Rename News to Market for watch module
Refactored the watch module by renaming News to Market. This includes renaming files, classes, and
 navigation destinations related to news to reflect the focus on market data.
