+++ app-watch/app/src/main/java/ir/composenews/MainViewModel.kt
@@ -2,7 +2,7 @@ package ir.composenews
 
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import ir.composenews.appwatch.navigation.graph.MainContract
+import ir.composenews.appwatch.navigation.MainContract
 import ir.composenews.base.BaseViewModel
 import ir.composenews.core_test.dispatcher.DispatcherProvider
 import ir.composenews.uimarket.model.MarketModel
